refactor(icons): consolidate icon registry and eliminate duplicate exports
