--- apps/web/src/components/ui/icons.ts
@@ -1,37 +1,0 @@
-export {
-    User,
-    MailOpen,
-    Phone,
-    MapPin,
-    Package,
-    ShoppingCart,
-    Bell,
-    Trash2,
-    AlertCircle,
-    Crown,
-    Settings as SettingsIcon,
-    FileText,
-    CheckCircle2,
-    Building2,
-    Save,
-    PlusCircle,
-    Eye,
-    Heart,
-    MessageSquare,
-    XCircle,
-    AlertTriangle,
-
-    LayoutDashboard,
-    ChevronRight,
-    LogOut,
-    Edit2,
-    CheckCheck,
-    EyeOff,
-    ArrowLeft,
-    CreditCard,
-    Clock,
-    Camera,
-    Upload,
-    ChevronLeft,
-    MoreHorizontal,
-} from "@esparex/ui";
